added helper files for common functions
